Refactor/Fixes/Features(registration)
