+++ nitorch/tools/registration/__init__.py
@@ -1,12 +1,11 @@
 from . import affine
+from . import joint
 from . import svf
 from . import svf_graph
+from . import smalldef
 from . import shoot
 from . import losses
+from . import objects
 from . import experimental_losses
 from . import phantoms
 from . import utils
-
-
-# cg   50x16 0.687523
-# ogm  222   0.687388
