+++ nitorch/tools/registration/shoot.py
@@ -1,8 +1,17 @@
+"""Diffeomorphic registration using Geodesic shooting.
+
+The function `shoot.register` is a prototype that requires input images
+to be defined on the same grid. See `registration.joint` for a more flexible
+model.
+
+The classes `RegisterStep` and `AutoRegStep` are implementation
+utilities. The main entry points are `register` and `autoreg`.
+"""
 from nitorch import spatial
-from nitorch.core import py, utils, linalg, math
+from nitorch.core import py, utils
 import torch
 from .utils import jg, jhj, defaults_velocity
-from . import plot as plt, optim as optm, losses, phantoms, utils as regutils
+from . import plot as plt, losses, phantoms, utils as regutils
 import functools
